requirements.txt
django==1.8
django-extensions
djangorestframework
requests
